Allow multiple ops.def and clean up code gen in general
Summary:
This is a cleanup and refactoring.
In its original form (changeset 6fdf915c057a) this diff caused a 5% regression
on ads CPU.  The root cause was an omission of link_whole = True, causing
symbols to be stripped in mode/opt and forcing the converter to fallback
causing patterns to be unmatched in the graph transform logic.  This version of
the diff tests for link_whole by including a C++ test of the transform

Reviewed By: yinghai

Differential Revision: D9040511

fbshipit-source-id: 3e19b89989aa68b021762d12af2d0b4111280b22